Update lambda build
